Porting project to a cleaner project standard.
Now using modern dotnet-core.
Fixed case of project files to be in line with common pascal-case standard used in dotnet.
Added new shared-items project as content project to be able to edit documentation in the IDE.
Changed all project types to be netstandard2.0 libraries, thanks to MonoGame supporting this build type now.