[Build] Added public and private section to commit message parser when generating changelogs
A commit message can now be split into a public part and a private part. The public part
content will be shown in the changelog. An optional private part is started after a commit
message line containing only "--". Any content after this marker will be ignored when generating
change logs.